move Input to components/legacy
