+++ src/components/CodeInput/CodeInput.js
@@ -23,7 +23,7 @@ import { TouchableWithoutFeedback } from 'react-native';
 import styled from 'styled-components/native';
 
 import { MediumText } from 'components/legacy/Typography';
-import TextInput from 'components/Input';
+import TextInput from 'components/legacy/Input';
 
 import { themedColors } from 'utils/themes';
 import { fontStyles, spacing } from 'utils/variables';
