+++ src/components/legacy/TextInput/TextInput.js
@@ -33,7 +33,7 @@ import Spinner from 'components/Spinner';
 import Icon from 'components/legacy/Icon';
 import Image from 'components/core/Image';
 import Button from 'components/legacy/Button';
-import Input from 'components/Input';
+import Input from 'components/legacy/Input';
 import ButtonText from 'components/legacy/ButtonText';
 import Tooltip from 'components/Tooltip';
